feat: Migrate local datastore from Isar to Drift
This commit migrates the local datastore implementation from Isar to Drift.
Key changes include:
- Replaced Isar-specific code with Drift database setup and operations.
- Updated data models and API to work with Drift tables.
- Refactored `_buildQuery` function into an extension for better organization.
- Updated dependencies and project documentation to reflect the change.
- Ensured all existing tests pass with the new Drift implementation.
