--- universal_datastores/assets/data/articles.json
@@ -1,10 +1,0 @@
-[
-  {
-    "title": "The Ultimate Guide to Flutter",
-    "content": "Flutter is a powerful open-source UI software development kit created by Google."
-  },
-  {
-    "title": "Getting Started with Isar",
-    "content": "Isar is a super fast, cross-platform, and easy-to-use database for Flutter."
-  }
-]
