+++ universal_datastores/lib/src/universal_datastores.dart
@@ -1,5 +1,5 @@
+/// The internal exports of the universal_datastores package.
+library;
+
 export 'article_datastore/article_datastore.dart';
-export 'article_datastore_drift/article_datastore_drift.dart';
-export 'database.dart';
 export 'datastore.dart';
-export 'exception.dart';
