+++ universal_datastores/lib/universal_datastores.dart
@@ -1,7 +1,7 @@
+/// {@template universal_datastores}
 /// A universal datastore for your Flutter applications.
+/// {@endtemplate}
 library;
 
 export 'src/article_datastore/article_datastore.dart';
-export 'src/database.dart';
 export 'src/datastore.dart';
-export 'src/exception.dart';
