+++ universal_datastores/memory-bank/productContext.md
@@ -4,7 +4,7 @@
 UI development is often blocked by the availability of backend APIs. This creates dependencies and slows down the development process. Frontend developers need a way to work with realistic data without waiting for the backend to be ready. Additionally, testing UI components with various data states can be cumbersome and time-consuming.
 
 ## Solution
-The `universal_datastore` package provides a local, Isar-backed datastore for mock data. This allows frontend developers to work with a stable, controllable, and fast local data source, decoupling them from the backend development lifecycle.
+The `universal_datastore` package provides a local, Drift-backed datastore for mock data. This allows frontend developers to work with a stable, controllable, and fast local data source, decoupling them from the backend development lifecycle.
 
 ## User Experience
 The package will provide a simple and intuitive API for developers to:
