+++ universal_datastores/memory-bank/techContext.md
@@ -3,20 +3,23 @@
 ## Technologies
 - **Dart:** The primary language for the package.
 - **Flutter:** The target framework for the example application.
-- **Isar:** A fast, cross-platform, and easy-to-use database for local data storage.
+- **Drift:** A reactive persistence library for Dart and Flutter.
 - **Mocktail:** A mocking library for unit tests.
 - **Very Good Analysis:** A set of linting rules for Dart and Flutter projects.
 
 ## Dependencies
-- `isar`
-- `isar_flutter_libs`
+- `drift`
 - `path_provider`
+- `path`
+- `sqlite3_flutter_libs`
 - `flutter` (for the example app)
 - `very_good_analysis` (dev dependency)
 - `mocktail` (dev dependency)
 - `test` (dev dependency)
+- `drift_dev` (dev dependency)
+- `build_runner` (dev dependency)
 
 ## Development Setup
 - A standard Flutter and Dart development environment.
-- The Isar generator will be used to generate the necessary code for the Isar database.
+- The Drift generator will be used to generate the necessary code for the Drift database.
 - The project will follow the Very Good Ventures engineering standards.
