+++ universal_datastores/pubspec.yaml
@@ -9,6 +9,8 @@ environment:
 dependencies:
   drift: ^2.18.0
   json_annotation: ^4.9.0
+  path: ^1.9.0
+  path_provider: ^2.1.3
   sqlite3_flutter_libs: ^0.5.24
 
 dev_dependencies:
@@ -18,4 +20,3 @@ dev_dependencies:
   mocktail: ^1.0.4
   test: ^1.25.8
   very_good_analysis: ^9.0.0
-
